requirements.txt
Flask
mysql-connector-python
python-dotenv
pandas
scikit-learn
requests
groq
nltk
